feat: seprate bytecode submissions
