+++ crates/assertion-da/da-server/Cargo.toml
@@ -8,7 +8,7 @@ name = "assertion-da"
 path = "bin/assertion-da.rs"
 
 [features]
-default = ["full-test"]
+default = ["full-test", "debug_assertions"]
 debug_assertions = []
 full-test = []
